Prepare for production launch

+++ resources/views/apikeys/index.blade.php
@@ -3,23 +3,23 @@
 @section('content')
 
 
-            <h2>Api Keys</h2>
+            <h2>Apps</h2>
 
             @if ( !$apikeys->count() )
-                You have no App Keys.
+                You have no Apps.
             @else
                 <table class="table">
                     <thead>
                         <tr>
                             <th>App Name</th>
-                            <th>API Key</th>
+                            <th>API Token</th>
                             <th>Actions</th>
                         </tr>
                     </thead>
                     <tbody>
                 @foreach( $apikeys as $apikey )
                     <tr>
-                        {!! Form::open(array('class' => 'form-inline', 'method' => 'DELETE', 'route' => array('apikeys.destroy', $apikey->slug))) !!}
+                        {!! Form::open(array('class' => 'form-inline', 'method' => 'DELETE', 'onsubmit' => 'return ConfirmDelete()', 'route' => array('apikeys.destroy', $apikey->slug))) !!}
                             <td><a href="{{ route('apikeys.show', $apikey->slug) }}">{{ $apikey->app_name }}</a></td>
                             <td>{{ $apikey->api_key }}</td>
                         <td>
@@ -33,9 +33,15 @@
                 </table>
             @endif
 
-            <p>
-                {!! link_to_route('apikeys.create', 'Create API Key', null, array('class' => 'btn btn-success')) !!}
-            </p>
+            @if ( $apikeys->count() < 5 )
 
+                <p>
+                    {!! link_to_route('apikeys.create', 'Create App', null, array('class' => 'btn btn-success')) !!}
+                </p>
+            @else
+
+                <p>You've reached the max apps allowed (5) for free users.</p>
+
+            @endif
 	
 @endsection
